+++ internal/terraform/default_tags.go
@@ -3,7 +3,6 @@ package terraform
 import (
 	"strings"
 
-	"github.com/hashicorp/hcl/v2"
 	"github.com/hashicorp/hcl/v2/hclsyntax"
 )
