+++ internal/terraform/helpers.go
@@ -27,3 +27,37 @@ func extractTraversalString(expr hcl.Expression, caseInsensitive bool) string {
 	}
 	return result
 }
+
+func mergeTags(tagMaps ...map[string]bool) map[string]bool {
+	merged := make(map[string]bool)
+	for _, m := range tagMaps {
+		for k, v := range m {
+			merged[k] = v
+		}
+	}
+	return merged
+}
+
+func filterMissingTags(requiredTags []string, effectiveTags map[string]bool, caseInsensitive bool) []string {
+	missing := []string{}
+	for _, tag := range requiredTags {
+		found := false
+		for existing := range effectiveTags {
+			if caseInsensitive {
+				if strings.EqualFold(existing, tag) {
+					found = true
+					break
+				}
+			} else {
+				if existing == tag {
+					found = true
+					break
+				}
+			}
+		}
+		if !found {
+			missing = append(missing, tag)
+		}
+	}
+	return missing
+}
