+++ internal/terraform/resources.go
@@ -3,7 +3,6 @@ package terraform
 import (
 	"strings"
 
-	"github.com/hashicorp/hcl/v2"
 	"github.com/hashicorp/hcl/v2/hclsyntax"
 )
 
@@ -83,28 +82,4 @@ func getResourceProvider(block *hclsyntax.Block, caseInsensitive bool) string {
 		defaultProvider = strings.ToLower(defaultProvider)
 	}
 	return defaultProvider
-}
-
-func filterMissingTags(requiredTags []string, effectiveTags map[string]bool, caseInsensitive bool) []string {
-	missing := []string{}
-	for _, tag := range requiredTags {
-		found := false
-		for existing := range effectiveTags {
-			if caseInsensitive {
-				if strings.EqualFold(existing, tag) {
-					found = true
-					break
-				}
-			} else {
-				if existing == tag {
-					found = true
-					break
-				}
-			}
-		}
-		if !found {
-			missing = append(missing, tag)
-		}
-	}
-	return missing
 }
