+++ internal/terraform/tags.go
@@ -23,16 +23,6 @@ func getTagMap(attr *hclsyntax.Attribute, caseInsensitive bool) (map[string]bool
 	return tags, nil
 }
 
-func mergeTags(tagMaps ...map[string]bool) map[string]bool {
-	merged := make(map[string]bool)
-	for _, m := range tagMaps {
-		for k, v := range m {
-			merged[k] = v
-		}
-	}
-	return merged
-}
-
 func findTags(block *hclsyntax.Block, caseInsensitive bool) (map[string]bool, error) {
 	if attr, ok := block.Body.Attributes["tags"]; ok {
 		return getTagMap(attr, caseInsensitive)
